# backend/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Role": {
				"enum": [
					0,
					1,
					2
				],
				"type": "number"
			},
			"Status": {
				"enum": [
					0,
					1,
					2,
					3
				],
				"type": "number"
			},
			"OnlineStatus": {
				"enum": [
					0,
					1
				],
				"type": "number"
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"onlineStatus": {
						"$ref": "#/components/schemas/OnlineStatus"
					},
					"statusTimeStamp": {
						"type": "string"
					},
					"logoutTime": {
						"type": "string"
					},
					"messages": {
						"items": {
							"$ref": "#/components/schemas/Message"
						},
						"type": "array"
					},
					"rooms": {
						"items": {
							"$ref": "#/components/schemas/Room"
						},
						"type": "array"
					},
					"historyStatus": {
						"items": {
							"$ref": "#/components/schemas/HistoryStatus"
						},
						"type": "array"
					},
					"accountStatus": {
						"$ref": "#/components/schemas/AccountStatus"
					}
				},
				"required": [
					"id",
					"username",
					"password",
					"role",
					"status",
					"onlineStatus",
					"statusTimeStamp",
					"logoutTime",
					"messages",
					"rooms",
					"historyStatus",
					"accountStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoomType": {
				"enum": [
					0,
					1,
					2,
					3,
					4
				],
				"type": "number"
			},
			"Message": {
				"properties": {
					"id": {
						"type": "string"
					},
					"sender": {
						"$ref": "#/components/schemas/User"
					},
					"content": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"room": {
						"$ref": "#/components/schemas/Room"
					}
				},
				"required": [
					"id",
					"sender",
					"content",
					"time",
					"status",
					"room"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Room": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/RoomType"
					},
					"messages": {
						"items": {
							"$ref": "#/components/schemas/Message"
						},
						"type": "array"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"type",
					"messages",
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HistoryStatus": {
				"properties": {
					"id": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"timeStamp": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"status",
					"timeStamp",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccountStatus": {
				"enum": [
					0,
					1
				],
				"type": "number"
			},
...
			"JoinRoomInput": {
				"properties": {
					"idList": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"idList"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateChatGroupInput": {
				"properties": {
					"roomId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/RoomType"
					}
				},
				"required": [
					"roomId",
					"userId",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateChatGroupInput": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"isJoin": {
						"type": "boolean"
					}
				},
				"required": [
					"userId",
					"isJoin"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "s23-ESN-SB1",
		"version": "0.0.1",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "SB1"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/rooms/{roomId}": {
			"get": {
				"operationId": "GetRoom",
				"responses": {
					"200": {
						"description": "Room",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Room"
								}
							}
						}
					}
				},
				"description": "Retrieve the room with all the related messages in the database",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roomId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateGroupChat",
				"responses": {
					"200": {
						"description": "Room",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Room"
								}
							}
						}
					}
				},
				"description": "Join or leave a group chat room",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roomId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateChatGroupInput"
							}
						}
					}
				}
			}
		},
		"/api/rooms": {
			"get": {
				"operationId": "GetChatGroup",
				"responses": {
					"200": {
						"description": "Rooms that are chatGroups",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Room"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieve all the chatGroups from the database",
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "JoinRoom",
				"responses": {
					"200": {
						"description": "Room",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Room"
								}
							}
						}
					}
				},
				"description": "Create and join a private chat room",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JoinRoomInput"
							}
						}
					}
				}
			}
		},
		"/api/rooms/chatgroup": {
			"post": {
				"operationId": "CreateGroupChat",
				"responses": {
					"200": {
						"description": "Room",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Room"
								}
							}
						}
					}
				},
				"description": "Create a group chat room",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateChatGroupInput"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}